Test case refactor
