+++ projects/subgraph-basin/tests/helpers/Liquidity.ts
@@ -1,3 +1,4 @@
+import { BigInt } from "@graphprotocol/graph-ts";
 import { Deposit, Withdraw } from "../../generated/schema";
 import { BEAN_ERC20, WETH } from "../../../subgraph-core/utils/Constants";
 import { handleAddLiquidity, handleRemoveLiquidity, handleRemoveLiquidityOneToken } from "../../src/WellHandler";
@@ -5,28 +6,28 @@ import { BEAN_SWAP_AMOUNT, SWAP_ACCOUNT, WELL, WELL_LP_AMOUNT, WETH_SWAP_AMOUNT 
 import { createContractCallMocks } from "./Functions";
 import { createAddLiquidityEvent, createRemoveLiquidityEvent, createRemoveLiquidityOneTokenEvent } from "./Well";
 
-export function createDefaultAddLiquidity(): string {
+export function mockAddLiquidity(tokenAmounts: BigInt[] = [BEAN_SWAP_AMOUNT, WETH_SWAP_AMOUNT]): string {
   createContractCallMocks();
-  let newEvent = createAddLiquidityEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, [BEAN_SWAP_AMOUNT, WETH_SWAP_AMOUNT]);
+  let newEvent = createAddLiquidityEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, tokenAmounts);
   handleAddLiquidity(newEvent);
   return newEvent.transaction.hash.toHexString() + "-" + newEvent.logIndex.toString();
 }
 
-export function createDefaultRemoveLiquidity(): string {
+export function mockRemoveLiquidity(tokenAmounts: BigInt[] = [BEAN_SWAP_AMOUNT, WETH_SWAP_AMOUNT]): string {
   createContractCallMocks();
-  let newEvent = createRemoveLiquidityEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, [BEAN_SWAP_AMOUNT, WETH_SWAP_AMOUNT]);
+  let newEvent = createRemoveLiquidityEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, tokenAmounts);
   handleRemoveLiquidity(newEvent);
   return newEvent.transaction.hash.toHexString() + "-" + newEvent.logIndex.toString();
 }
 
-export function createRemoveLiquidityOneBean(): string {
+export function mockRemoveLiquidityOneBean(): string {
   createContractCallMocks();
   let newEvent = createRemoveLiquidityOneTokenEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, BEAN_ERC20, BEAN_SWAP_AMOUNT);
   handleRemoveLiquidityOneToken(newEvent);
   return newEvent.transaction.hash.toHexString() + "-" + newEvent.logIndex.toString();
 }
 
-export function createRemoveLiquidityOneWeth(): string {
+export function mockRemoveLiquidityOneWeth(): string {
   createContractCallMocks();
   let newEvent = createRemoveLiquidityOneTokenEvent(WELL, SWAP_ACCOUNT, WELL_LP_AMOUNT, WETH, WETH_SWAP_AMOUNT);
   handleRemoveLiquidityOneToken(newEvent);
